fix(core): split pre-spawn view rules module
